data-collection/requirements.txt
python-dotenv~=1.0.1
requests~=2.32.3
pandas~=2.2.3
psycopg2-binary~=2.9.10
SQLAlchemy~=2.0.36
GeoAlchemy2~=0.15.2
numpy~=2.0.2